Interface Users & User shared type
