+++ server/serverTypes.ts
@@ -14,12 +14,3 @@ export interface PlayerSocket extends SocketIO.Socket {
 export interface Enemies {
   [key: string]: Enemy;
 }
-
-export interface Users {
-  [key: string]: User;
-}
-
-interface User {
-  username: string;
-  role: string;
-}
